Update case of binary in goreleaser
We hit an issue when attempting to use this plugin as installed from the
GitHub release archive. It appears that Kustomize expects the case of
the plugin's binary to match the `Kind` in the hashannotator object.
This change updates the goreleaser.yaml file to create a binary with the
correct tense.